+++ test/helpers/mock-servers/service-testers/stream-fetchers/implementation/read-rows-fetcher.ts
@@ -16,7 +16,7 @@
 // ** https://github.com/googleapis/gapic-generator-typescript **
 // ** All changes to this file may be overwritten. **
 
-import {GetRowsOptions, Table} from '../../../../../table';
+import {GetRowsOptions, Table} from '../../../../../../src/table';
 import internal = require('stream');
 import {StreamFetcher} from '../stream-fetcher';
